Evals the return of CFBundleShortVersionString
If the value in CFBundleShortVersionString is a link to another variable (like “$(MARKETING_VERSION)”), the parsing of the variable was when seting the BUILDNUMBER.

In that case, FIRSTCOMPONENT was set to ‘$(MARKETING_VERSION)’, SECONDCOMPONENT was set to ‘’ and THIRDCOMPONENT was set to ‘’ (and thus reset to 0).  This lead to having a BUILDNUMBER set to $(MARKETING_VERSION)..0.$(DATE).

With two dots following each other, upload of the build would be denied by Apple.